feat: Add Activity Metrics Service and multi-thread architecture
- Add Activity Metrics Service (4fd5b690-...) for real-time running metrics
- Move Total Step Count and Activity Step Count from Information Service to Activity Metrics Service
- Add Secondary Device Service (4fd5b6a0-...) for consolidated secondary device info
- Implement packed data structures for efficient BLE transmission:
  - Packed Device Status (16 bytes combining multiple status fields)
  - Packed File Notification (64 bytes combining file ID and path)
- Create multi-thread architecture for activity metrics processing:
  - sensor_data module (100Hz sensor processing)
  - realtime_metrics module (real-time metrics calculation with 1Hz BLE updates)
  - analytics module (complex analytics and baseline tracking)
- Update Bluetooth module to integrate with new architecture
- Add comprehensive documentation for all changes
- Update all document dates to July 2025 format
- Add .clang* files to .gitignore

BREAKING CHANGE: Step count characteristics moved from Information Service to Activity Metrics Service. Mobile apps need to update their service discovery.

+++ src/bluetooth/ble_d2d_rx_client.cpp
@@ -1081,7 +1081,7 @@ static uint8_t weight_measurement_notify_handler(struct bt_conn *conn,
     uint16_t weight_kg_x10 = *(const uint16_t *)data;
     float weight_kg = weight_kg_x10 / 10.0f;
     LOG_INF("=== RECEIVED WEIGHT MEASUREMENT FROM SECONDARY ===");
-    LOG_INF("  Weight: %.1f kg (raw=%u)", weight_kg, weight_kg_x10);
+    LOG_INF("  Weight: %.1f kg (raw=%u)", (double)weight_kg, weight_kg_x10);
 
     // Forward to information service for aggregation with primary weight
     // The primary device should aggregate both feet weights
